Added local access ability
You can fill in your server IP, Port, and token in the system.py and piplaystate.py script to use the local access method instead of the cloud access method. This will become the preferred method in the near future. Options to add and places in the DB will appear for these soon.