multiple adds and a speed fix
Changed script so it only logins into your plex server when it needs to instead of every time it runs. Added option for "whatsafterthat" Used for seeing the 2nd item in the queue. If no second item, a second item will be autopopulated following the Agent Smith content add rules. Used the API to add ability to skipback and skipahead (forward 30 seconds/back 15ish?). Rementioning the fix for whatupnext it was fixed in a previous version posted shortly before this one.