list search and binge watch
In many of the lists that are returned, where given the Yes/No prompt you can instead enter "letter  + a"(replace a with desired letter) and the list will jump to items that start with that letter. 
A binge watch mode has been added. This will repeat the specified show/movie until a new mode is selected. usage example- binge."movie.The Way of the Gun" OR binge.Psych