used for ts TV add issues
used for troubleshooting TV add issues. It will output text and try to write to a file when there is an error. If you use this to report an issue, please include that text file, as well as what the metadata from your plex library has for what should be where "ERRORHERE" is. This usually is caused by an unsupported character issue that I don't know about and need to figure out how to properly fix. 

NOTE: You will need to manually fix the home dir link on the first line of the file to use this one. Place it where your upddateddb_pi.py file is, and use the same path that is set in that file.