commercial/preroll support
can use the following command to add either a commercial or preroll to a block:
addtoblock "blockname" playcommercial/preroll

can test commercial playback by using "playcommercial" command.

Note: before using you need to run the latest upddatedb_pi.py file with "getcommercials" and/or "getprerolls" to populate/update those tables. You can have multiple commercials and prerolls. If it were not obvious, you are not limited to only using commercials here. I have jokes and random movie clips in my commercials folder. A random one will be used when that option is called. It is recommended you use "reorderblock 'blockname'" after adding a commercial or preroll to a block so you can reorder it as desired.  

Note2: This should be considered a WIP in progress feature. In case I did not catch all the potential failures, If you find any failures please open a bug report and include the output from a "startnextprogram" with the block in question, as well as a "explainblock" for that