fixing apostrophe support
You may need to update your tbnplex db to use the fixes.
note: unmatched items in your plex db may cause the update db script to error. 
note2: the displaying of the double apostrophe in outputs is not a bug and is intentional for the time being for debug purposes.