start of the crappy UI
this is the start of the crappy ui. You can drop this in your apaches web directory and use them as a UI. I recommend better people make their own. I'd consider this an experimental feature that will have limited development action. The query DB option only supports a few commands currently. It may be expanded. Feel free to expand it and add suggestions with code for changes. 
all 3 .php files are needed to make use of the limited features available.