replace show in block option
Added ability to have a show automaticlly replaced in a block when it has reached the end of the episode list. Commands:
replaceshowinblock showtoreplace newshow blockaffected playstatus --> associates the new show with the replaced show and block.
replacestatus show --> checks if the specified show has a replacement setup
replacecheck --> does the replace action. used for testing

Note: if the show is played outside the block and reaches the end of the episode list, this will also trigger the replace action to happen.