updatedb for specific show
This script can be used when you want to update the TBN-Plex database for a specific show. This script will not delete and rebuild the TVshowlist and shows tables. Instead it looks for the show specified and adds entries for those that do not already exist.
Note: It will not update errors in exsting content. You will still need to use the updatedb script and do a updatetv for that. 
Usage: python getshow.py "show name here"