queuetoplaylist updates
updated system script so that when a block package is active, and queuetoplaylist is enabled, the blocks details will be pushed to the plexplaylist in place of the queue, while the block is active. 
Notes: If more than one random item is in a block, only the random up next is pushed, with the others skipped, IE not reflected in your plex playlist. This is because the system script only determines the upnext random, and not the ones after that until needed. 
Commercials and prerolls will get added to the playlist. If you are using random, those reflected in the playlist may not be the ones that play via the startnextprogram action.