version 3.00a files
New version removes "shows" table and instead uses api calls to get everything it was used for previously. Note: It is recommended you first make a backup of your myplex.db before updating to this version.
skipahead and skipback improvements. Now you can use "xm" where "x" is the number of minutes to skip ahead, or "xs" where "x" is the number of seconds to skip ahead. Defaults to 30 seconds.
help flag "-h" - Approximatly 50% of the commands now support the "-h" flag, where you can get a little howto and description for the command. Update the helptable with the "updatehelp" command. The online help file this table pulls from is being updated to have details for all commands, so remember to updatehelp peridioclly if you use this option. 
See comments for further details.